pyproject.toml
[tool.poetry]
name = "sophon"
version = "0.1.0"
description = "Django project for dataset collection"
authors = ["Stefano Pigozzi <256895@studenti.unimore.it>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
psycopg2 = "^2.8.6"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
pandaSDMX = "^1.4.2"
pydantic = "~1.7.3"
cryptography = "^3.4.7"
drf-jwt = "^1.19.0"
PyJWT = "^2.0.1"
django-cors-headers = "^3.7.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
